Simplify search provider interface
The search provider logic was unneccesarily complicated by trying
to enforce a common interface for the formatResults function. This
also coupled the API response callback function signature of the
search providers, which will cause problems if a provider needs a
different callback.

By just calling the provider and letting it handle the result
formatting this, both the documentation as well as the elasticsearch
provider code are simplified significantly.