Add regulator service with unjudged-only distribution of work
The regulator service sits between the search index and provider,
and filters the results returned by the provider with the set of
bookmarks, so only unjudged results are returned. It will return
the same number of results (if available) by re-issuing requests
to the provider until there are enough results.