Add support for unbookmarked soft distribution of labour
When unbookmarked soft distribution labour is enabled, all fetched
results including bookmarked results are returned to the frontend.
How to mark bookmarked results is left up to the frontend.
However, the same number of unbookmarked results are always
returned to the frontend, since those are the results the user
can bookmark.