Add indri search provider
The indri search provider uses node-indri to provide full-text
search with support for relevance feedback using the indri search
engine.

Node-indri currently needs to be copied to the lib folder for
testing (not included in the repo to prevent polluting it with
unnecessary data). It will later be included as a module.