Add missing fields and decouple text result format from indri
Missing name, snippet or text fields are now replaced with a
placeholder text when missing.

The new formatfunction in the indri provider is partially redundant
with the node-indri's includeFields, but adding this mapping allows
us to decouple the result format from indri specifics (such as
having metaFields returned in a separate field).