Show non-viewed results from lower page numbers on current page
By keeping track of the results the user has viewed, we can show
any results that are on previous pages, but the user has not viewed
yet. This situation can arise when relevance feedback promotes
results that used to be on a higher numbered page to a lower
numbered page, while the user already visited that page before
the results were promoted.

This feature makes it impossible for the user to directly navigate
to a later page, since it will always show the highest ranked results
that the user has not viewed yet. However, this is not a useful
navigational action during normal search sessions, so that is
an acceptable tradeoff for showing the user the most relevant results
during a normal search session where they start at the first page
and navigate to later pages.