Page numbers start after a fixed number of uncollapsible results
When the user views a page, we always show 10 uncollapsible results
(this was already the case). This means that the total number of
results shown on the page can vary. Therefore we redefine the start
of a page to not be after a fixed number of results, but to be after
a fixed number of uncollapsible results. This prevents overlap
between pages when using distribution of labour.