Track last page position to prevent skipping results
When the user navigates from one page to a subsequent page while
bookmarks or excludes have been added in the meantime, we do not
want the page position for a user to jump, because they would skip
potentially relevant results. Tracking the position of the last
element of the last page, and using it to start the next page
fixes this.