Fix socket channels by joining after group is assigned
Sockets were being joined into channels when a new user joins the
system, while their sessionId later changes when a group is assigned.
This issue was masked by the fact that the sessionId was not set
properly, causing it to be undefined, and syncData updates to be
sent to all users (instead of just within a session), meaning that
multiple simulataneous sessions were affecting eachother.

This also replaces the sessionId with groupId, since we now use
multiple sessionId's when one group executes subsequent topics.
Although currently syncData is only used on group formation, this
could lead to bugs later on.